rename wcs_tools.py to tools.py
