+++ ndcube/wcs/tests/test_wcs_tools.py
@@ -4,7 +4,7 @@ from astropy.wcs import WCS
 from astropy.wcs.wcsapi import SlicedLowLevelWCS
 from numpy.testing import assert_array_almost_equal, assert_array_equal
 
-from ndcube.wcs.wcs_tools import unwrap_wcs_to_fitswcs
+from ndcube.wcs.tools import unwrap_wcs_to_fitswcs
 from ndcube.wcs.wrappers import ResampledLowLevelWCS
